refactor : PathFinder 인테페이스 선언 및 구현 처리
