+++ src/main/java/nextstep/subway/path/application/PathService.java
@@ -1,6 +1,7 @@
 package nextstep.subway.path.application;
 
 import nextstep.subway.line.application.LineService;
+import nextstep.subway.path.domain.DijkstraPathFinder;
 import nextstep.subway.path.domain.PathFinder;
 import nextstep.subway.path.domain.ShortestPath;
 import nextstep.subway.path.dto.PathResponse;
@@ -24,7 +25,7 @@ public class PathService {
     public PathResponse findShortestPath(Long sourceStationId, Long targetStationId) {
         Station sourceStation = stationService.findById(sourceStationId);
         Station targetStation = stationService.findById(targetStationId);
-        PathFinder pathFinder = new PathFinder(lineService.findAllLines());
+        PathFinder pathFinder = new DijkstraPathFinder(lineService.findAllLines());
 
         ShortestPath shortestPath = pathFinder.findShortestPath(sourceStation, targetStation);
